PythonFlaskTemplate/requirements.txt
Flask~=1.1.2
ruamel.yaml~=0.16.12
